src/SmartConfig.LangGraph/src/utils/config.py
from __future__ import annotations

import os
from dataclasses import dataclass
from typing import Optional

from dotenv import load_dotenv


load_dotenv()


@dataclass(frozen=True)
class Settings:
    openai_api_key: Optional[str] = os.getenv("OPENAI_API_KEY")
    tavily_api_key: Optional[str] = os.getenv("TAVILY_API_KEY")


settings = Settings()


