src/SmartConfig.LangGraph/requirements.txt
langgraph>=0.6.0
langchain>=0.2.16
langchain-openai>=0.1.22
langchain-tavily>=0.0.2
python-dotenv>=1.0.1
pydantic>=2.8.2
uvicorn>=0.30.3
fastapi>=0.112.0
tavily-python>=0.5.0
beautifulsoup4>=4.12.3
httpx>=0.27.0
